Removes implementation changes, keeps only specs
