--- lib/superform/field.rb
@@ -1,36 +1,0 @@
-module Superform
-  class Field < Node
-    attr_reader :dom
-
-    def initialize(key, parent:, object: nil, value: nil)
-      super key, parent: parent
-      @object = object
-      @value = value
-      @dom = Superform::DOM.new(field: self)
-    end
-
-    def value
-      if @object and @object.respond_to? @key
-        @object.send @key
-      else
-        @value
-      end
-    end
-    alias :serialize :value
-
-    def assign(value)
-      if @object and @object.respond_to? "#{@key}="
-        @object.send "#{@key}=", value
-      else
-        @value = value
-      end
-    end
-    alias :value= :assign
-
-    # Wraps a field that's an array of values with a bunch of fields
-    # that are indexed with the array's index.
-    def collection(&)
-      @collection ||= FieldCollection.new(field: self, &)
-    end
-  end
-end
